Move closeAll into its own module file
