+++ packages/modal/src/js/index.js
@@ -6,6 +6,7 @@ import { register } from './register';
 import { deregister } from './deregister';
 import { open } from './open';
 import { close } from './close';
+import { closeAll } from './closeAll';
 import { replace } from './replace';
 import { getModalID, getModalElements } from './helpers';
 
@@ -87,19 +88,11 @@ export default class Modal extends Collection {
     return close.call(this, modal, transition);
   }
 
-  async closeAll(exclude = false, transition = this.settings.transition) {
-    const result = [];
-    await Promise.all(this.stack.map(async (modal) => {
-      if (exclude && exclude === modal.id) {
-        Promise.resolve(); 
-      } else {
-        result.push(await modal.close(transition));
-      }
-    }));
-    return result;
+  closeAll(exclude = false, transition = this.settings.transition) {
+    return closeAll(this.stack, exclude, transition);
   }
 
-  async replace(id, transition = this.settings.transition) {
+  replace(id, transition = this.settings.transition) {
     const modal = this.get(id);
     if (!modal) return false;
     return replace.call(this, modal, transition);
